Vehicle add endpoint implementation
I added the vehicle add endpoint, and, when an vehicle is created, it will have the default configurations for the recomandations reminder system

+++ CarTrack_API/EntityLayer/Dtos/VehicleDto/VehicleResponseDto.cs
@@ -4,7 +4,7 @@ public class VehicleResponseDto
 {
     public int Id { get; set; }
     public string Vin { get; set; }
-    public int Mileage { get; set; }
+    public double Mileage { get; set; }
     public required string ModelName { get; set; }
     public int Year { get; set; }
 }
